refactor: apply the snake case convention for the file names of llm_client modules
